+++ src/ShiftedClock.php
@@ -20,6 +20,11 @@ final class ShiftedClock implements ClockInterface
         $this->timeZone = $timeZone;
     }
 
+    public static function fromDateString(string $dateInterval, DateTimeZone $timeZone = null): self
+    {
+        return new self(DateInterval::createFromDateString($dateInterval), $timeZone);
+    }
+
     public function now(): DateTimeImmutable
     {
         return (new DateTimeImmutable('now', $this->timeZone))->add($this->dateInterval);